refactor: 프로젝트 구조 대폭 개선 및 전문적인 Python 패키지 구조로 전환
## 주요 변경사항

### 🏗️ 프로젝트 구조 재정리
- 표준 Python src/ 레이아웃 적용
- 소스코드를 src/ 디렉토리로 이동 및 체계화
- 데이터와 코드 완전 분리 (data/, logs/ 디렉토리)

### 📁 디렉토리 구조 개선
- src/ - 메인 소스코드 (config, core, models, training, utils)
- cli/ - CLI 인터페이스 스크립트
- data/ - 사용자 데이터 (generated_images, user_loras, DB)
- logs/ - 로그 파일
- tests/ - 테스트 코드

### 🔧 설정 시스템 개선
- config.py → src/config/ 모듈로 분리
- settings.py와 logging_config.py로 관심사 분리
- 환경변수 지원으로 설정 유연성 향상

### 📦 파일명 정리
- 'improved_' 접두사 제거로 명확한 네이밍
- improved_evaluators.py → evaluators.py
- improved_diversity_evaluator.py → diversity_evaluator.py
- improved_trainer.py → ddpo_trainer.py

### 🛠️ CLI 구조 개선
- 메인 CLI와 피드백 CLI 분리
- feedback.py 독립 실행 스크립트 추가
- main.py를 엔트리 포인트로 단순화

### 📚 프로젝트 메타데이터 추가
- pyproject.toml 추가 (현대적 Python 패키지 설정)
- requirements-dev.txt 추가 (개발 의존성 분리)
- .gitignore 업데이트 (data/ 디렉토리 제외)

### 🔗 Import 경로 수정
- 모든 소스코드의 import 경로 src/ 구조에 맞게 수정
- 상대/절대 import 정리로 구조 일관성 확보

### 📖 문서 업데이트
- README.md에 프로젝트 구조 섹션 추가
- 새로운 디렉토리 구조와 사용법 반영
