+++ cli/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python3
+"""
+CLI 모듈 - 명령줄 인터페이스
+"""
+
+from .main import main
+from .feedback_cli import GentleFeedbackCLI
+
+__all__ = ["main", "GentleFeedbackCLI"]
