+++ pyproject.toml
@@ -1,0 +1,80 @@
+[build-system]
+requires = ["setuptools>=45", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "emoseum"
+version = "0.1.0"
+description = "감정 기반 디지털 치료 이미지 생성 시스템"
+readme = "README.md"
+requires-python = ">=3.8"
+license = {text = "MIT"}
+authors = [
+    {name = "Emoseum Team"}
+]
+keywords = ["emotion", "therapy", "image-generation", "ai", "mental-health"]
+classifiers = [
+    "Development Status :: 3 - Alpha",
+    "Intended Audience :: Healthcare Industry",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Topic :: Scientific/Engineering :: Medical Science Apps.",
+]
+dependencies = [
+    "torch>=2.0.0,<3.0.0",
+    "torchvision>=0.15.0,<1.0.0",
+    "numpy>=1.21.0,<2.0.0",
+    "Pillow>=9.0.0,<11.0.0",
+    "transformers>=4.30.0,<5.0.0",
+    "diffusers>=0.20.0,<1.0.0",
+    "accelerate>=0.20.0,<1.0.0",
+    "opencv-python>=4.7.0,<5.0.0",
+    "datasets>=2.10.0,<3.0.0",
+    "peft>=0.4.0,<1.0.0",
+    "tqdm>=4.64.0,<5.0.0",
+    "wandb>=0.15.0,<1.0.0",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.0.0,<8.0.0",
+    "pytest-cov>=4.0.0,<5.0.0",
+    "pytest-mock>=3.10.0,<4.0.0",
+    "black>=23.0.0,<24.0.0",
+    "flake8>=6.0.0,<7.0.0",
+    "isort>=5.12.0,<6.0.0",
+    "mypy>=1.0.0,<2.0.0",
+    "sphinx>=5.0.0,<6.0.0",
+    "sphinx-rtd-theme>=1.2.0,<2.0.0",
+    "pre-commit>=3.0.0,<4.0.0",
+]
+
+[project.scripts]
+emoseum = "main:main"
+emoseum-feedback = "feedback:main"
+
+[tool.black]
+line-length = 88
+target-version = ['py38']
+
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+line_length = 88
+
+[tool.mypy]
+python_version = "3.8"
+warn_return_any = true
+warn_unused_configs = true
+disallow_untyped_defs = true
+ignore_missing_imports = true
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+addopts = "-v --cov=. --cov-report=html --cov-report=term-missing"
