+++ requirements-dev.txt
@@ -1,0 +1,20 @@
+# Development dependencies
+-r requirements.txt
+
+# Testing
+pytest>=7.0.0,<8.0.0
+pytest-cov>=4.0.0,<5.0.0
+pytest-mock>=3.10.0,<4.0.0
+
+# Code quality
+black>=23.0.0,<24.0.0
+flake8>=6.0.0,<7.0.0
+isort>=5.12.0,<6.0.0
+mypy>=1.0.0,<2.0.0
+
+# Documentation
+sphinx>=5.0.0,<6.0.0
+sphinx-rtd-theme>=1.2.0,<2.0.0
+
+# Pre-commit hooks
+pre-commit>=3.0.0,<4.0.0
