+++ src/__init__.py
@@ -1,0 +1,17 @@
+#!/usr/bin/env python3
+"""
+Emoseum - 감정 기반 디지털 치료 이미지 생성 시스템
+"""
+
+__version__ = "0.1.0"
+__author__ = "Emoseum Team"
+
+from .core.therapy_system import EmotionalImageTherapySystem
+from .models.emotion import EmotionEmbedding
+from .models.user_profile import UserEmotionProfile
+
+__all__ = [
+    "EmotionalImageTherapySystem",
+    "EmotionEmbedding", 
+    "UserEmotionProfile"
+]
