+++ src/core/__init__.py
@@ -3,6 +3,6 @@
 Emoseum Core - 핵심 시스템 모듈
 """
 
-from core.therapy_system import EmotionalImageTherapySystem
+from .therapy_system import EmotionalImageTherapySystem
 
 __all__ = ["EmotionalImageTherapySystem"]
