+++ src/models/__init__.py
@@ -3,16 +3,16 @@
 Emoseum Models - 감정 분석 및 사용자 프로필 모델
 """
 
-from models.emotion import EmotionEmbedding
-from models.emotion_mapper import AdvancedEmotionMapper
-from models.user_profile import UserEmotionProfile
-from models.lora_manager import PersonalizedLoRAManager
-from models.improved_evaluators import (
+from .emotion import EmotionEmbedding
+from .emotion_mapper import AdvancedEmotionMapper
+from .user_profile import UserEmotionProfile
+from .lora_manager import PersonalizedLoRAManager
+from .evaluators import (
     ImprovedAestheticEvaluator,
     ImprovedEmotionEvaluator,
     ImprovedPersonalizationEvaluator
 )
-from models.improved_diversity_evaluator import ImprovedDiversityEvaluator
+from .diversity_evaluator import ImprovedDiversityEvaluator
 
 __all__ = [
     "EmotionEmbedding",
